+++ tests/codegen/fail/interface/derive_wrong_item.rs
@@ -1,0 +1,12 @@
+use juniper::{GraphQLInterface, GraphQLObject};
+
+#[derive(GraphQLObject)]
+pub struct ObjA {
+    test: String,
+}
+
+#[derive(GraphQLInterface)]
+#[graphql(for = ObjA)]
+enum Character {}
+
+fn main() {}